revert: rapidadapter size_t

+++ src/gamelogic/rapidXML/RapidXMLAdapter.cpp
@@ -135,9 +135,9 @@ namespace GameLogic {
     /// This function returns the objects that is used in this .tmx.
     /// RapidXMLAdapter::setup_document has to be called before using this
     /// This should be called as the last step of parsing the .tmx
-    vector<tuple<int, int, char*>> RapidXMLAdapter::get_objects()
+    vector<tuple<size_t, size_t, char*>> RapidXMLAdapter::get_objects()
     {
-        vector<tuple<int, int, char*>> object_list;
+        vector<tuple<size_t, size_t, char*>> object_list;
         for ( xml_node<> * object_node = this->_object_group_node->first_node( "object" ); object_node; object_node = object_node->next_sibling() ) {
             int x = std::stoi( object_node->first_attribute( "x" )->value() ) / this->_tile_width;
             int y = std::stoi( object_node->first_attribute( "y" )->value() ) / this->_tile_height;
